requirements.txt
absl-py
aioice
aiortc
altair
appdirs
astunparse
async-generator
attrs
av
beautifulsoup4
blinker
bs4
CacheControl
cachetools
certifi
cffi
charset-normalizer
click
cryptography
cssselect
cycler
decorator
dnspython
entrypoints
exceptiongroup
fake-useragent
firebase-admin
flatbuffers
fonttools
gast
gitdb
GitPython
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-pasta
google-resumable-media
googleapis-common-protos
grpcio
grpcio-status
h11
h5py
httplib2
idna
ifaddr
importlib-metadata
importlib-resources
jax
Jinja2
jsonschema
keras
kiwisolver
libclang
lxml
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdurl
mediapipe
ml-dtypes
msgpack
numpy
oauthlib
opencv-contrib-python
opencv-python-headless
opt-einsum
outcome
packaging
pandas
parse
Pillow
proto-plus
protobuf
pyarrow
pyasn1
pyasn1-modules
PyAudio
pycparser
pydeck
pyee
Pygments
PyJWT
pylibsrtp
pymongo
Pympler
pyOpenSSL
pyparsing
pyppeteer
pyquery
pyrsistent
PySocks
python-dateutil
python-decouple
pytz
pytz-deprecation-shim
requests
requests-html
requests-oauthlib
rich
rsa
scipy
selenium
setuptools-scm
six
smmap
sniffio
sortedcontainers
soupsieve
SpeechRecognition
streamlit
streamlit-webrtc
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
tensorflow-io-gcs-filesystem
termcolor
toml
tomli
toolz
tornado
tqdm
trio
trio-websocket
typing_extensions
tzdata
tzlocal
uritemplate
urllib3
validators
w3lib
websockets
Werkzeug
wrapt
wsproto
zipp
